+++ src/rx-collection-helper.ts
@@ -38,11 +38,13 @@ export function fillObjectDataBeforeInsert<RxDocType>(
 ): RxDocumentData<RxDocType> {
     data = flatClone(data);
     data = fillObjectWithDefaults(schema, data);
-    data = fillPrimaryKey(
-        schema.primaryPath,
-        schema.jsonSchema,
-        data
-    );
+    if (typeof schema.jsonSchema.primaryKey !== 'string') {
+        data = fillPrimaryKey(
+            schema.primaryPath,
+            schema.jsonSchema,
+            data
+        );
+    }
     data._meta = getDefaultRxDocumentMeta();
     if (!Object.prototype.hasOwnProperty.call(data, '_deleted')) {
         data._deleted = false;